deepseek_v4_v100: V100 fp16 single-request port of V4-Flash (W4A16 AutoRound)
Wires DeepSeek-V4-Flash (Intel/DeepSeek-V4-Flash-W4A16-AutoRound, 290B
total / ~37B activated, 256 experts, MLA + sparse attention + Hyper-
Connections) to run on 8x V100 SXM2 32GB. Builds on the kernel ports
(8ac0e387f) and the V100 attention layer math (6e75ed19d) added earlier
on this branch.

What this adds:

  - vllm/model_executor/models/deepseek_v4.py
      DeepseekV4ForCausalLM, DeepseekV4Model, DeepseekV4DecoderLayer,
      DeepseekV4MoE, V4Attention (vLLM AttentionLayerBase), pure-pytorch
      Hyper-Connections helpers (_hc_pre / _hc_post / _hc_head), custom
      load_weights with W4A16 dequant pre-processor and FusedMoE expert
      mapping, custom routing function for sqrtsoftplus + noaux_tc topk +
      hash-MoE for the first num_hash_layers layers.

  - vllm/v1/attention/backends/deepseek_v4_v100.py
      DeepSeekV4FlashV100Backend / MetadataBuilder / Impl. Strict-V100
      (cc=(7,0)), fp16-only, registered in the AttentionBackendEnum as
      DEEPSEEK_V4_FLASH_V100. Backend exists for backend-test coverage;
      V4Attention.forward currently calls v100 sparse_attn directly to
      keep the single-request path simple.

  - vllm/transformers_utils/configs/deepseek_v4.py
      DeepseekV4Config (model_type="deepseek_v4"), wired into
      _CONFIG_REGISTRY and configs __init__.

  - vllm/model_executor/models/registry.py
      DeepseekV4ForCausalLM entry mapping to deepseek_v4.

  - vllm/model_executor/layers/deepseek_v4_v100_attention.py (small fix)
      V4Compressor's wkv/wgate now have explicit dtype=float32 (the
      reference casts x to fp32 before applying these linears, so the
      linears must accept fp32 inputs). Without this they inherited
      torch.get_default_dtype() and broke depending on caller context.

The non-obvious bit -- _hc_post fp16 clamp:

  Layer 0 hash-MoE routes BOS to a specific expert combination that
  produces a ~14x larger MoE output than for content tokens. Combined
  with the standard ~1.2x per-layer residual growth, pos 0 hits fp16
  max (65504) by layer 27, becomes inf at layer 28, and at the next
  ratio=4 layer V4Compressor's RMSNorm produces NaN (rsqrt(inf)=0,
  inf*0=NaN). NaN propagates to every position, LM-head logits become
  NaN, greedy argmax over NaN deterministically picks token 0 (BOS),
  and decode degenerates to BOS spam.

  The reference inference impl runs in bf16 (range ~3e38) so this
  never hits. We can't switch to bf16 (no native bf16 mma on V100).

  Fix: clamp the fp32 intermediate inside _hc_post to +-50000 BEFORE
  casting to fp16. Healthy magnitudes stay <1000 so the clamp is a
  no-op for raw prompts; only activates when pos 0 grows toward fp16
  max. Throughput unchanged at 5.15 tok/s.

Constraints on this PR (deferred for future work):

  - Single-request only. V4Attention.forward asserts bsz==1 and uses
    module-level KV buffers on V4Compressor / V4Indexer. Multi-request
    requires lifting the windowed/compressed/indexer KV into vLLM's
    paged cache manager (mirroring upstream's DeepseekV4SWACache /
    DeepseekV4IndexerCache pattern).

  - TP=8 is the tested topology (n_groups=8, n_heads=64). TP<8 paths
    work mathematically (per-group einsum for wo_a) but aren't
    exercised; TP=8 collapses wo_a to a quantized ColumnParallelLinear
    via the single-group-per-rank trick.

  - --enforce-eager required (cudagraph not yet wired). Three engine
    knobs needed for V4-Flash on V100: max_num_seqs=4 (sampler warmup
    OOM headroom), enable_prefix_caching=False (only the No-Prefix
    coordinator tolerates the model's empty KVCacheSpec), and the
    backend reports get_kv_cache_spec=None so vLLM treats the model as
    attention-free for KV pool sizing.

  - MTP layers are skipped at load time (skip_substrs=["mtp."]).

Validation:

  - tests/models/test_deepseek_v4_v100_tp8_long_chat.py: 73-token
    chat-formatted poem prompt produces a 146-token ABAB-rhymed poem
    about GPUs, finishes at EOS naturally, 0/146 BOS, 5.15 tok/s.
    Model correctly interprets <|User|>/<|Assistant|>/<think> tokens
    and follows the system-prompt directive ("never include any
    preamble outside the poem itself").

  - tests/models/test_deepseek_v4_v100_tp8_bisect.py: regression test
    for the chat-token degeneracy. Four prompt shapes (3-tok raw,
    12-tok raw, 30-tok raw, BOS+chat-special) all produce coherent
    output post-fix, 0 BOS each.

  - tests/kernels/attention/test_deepseek_v4_v100_attention_equivalence.py:
    per-token L2 rel_err < 0.6% (mean) vs the reference's bf16
    Attention.forward at single-rank for ratios 0/4/128.

  - Other tests cover backend contract, model class instantiation,
    weight loader for the first 5 shards, synthetic-config forward,
    TP=8 instantiation, and TP=8 serve smoke.

Run from the repo root:

    PATH=$PATH:/usr/local/cuda-12.8/bin /home/admin/venv/bin/python \
        tests/models/test_deepseek_v4_v100_tp8_long_chat.py

(repo's tests/conftest.py pulls in vllm._C which the wheel install
doesn't ship; tests have explicit __main__ entries to bypass pytest.)

+++ vllm/model_executor/layers/deepseek_v4_v100_attention.py
@@ -212,11 +212,16 @@ class V4Compressor(nn.Module):
         self.ape = nn.Parameter(
             torch.empty(compress_ratio, coff * self.head_dim, dtype=torch.float32)
         )
-        # In the checkpoint wkv/wgate are bf16 weights; we cast to fp16 at
-        # load time. The reference keeps these in fp32 for "convenience";
-        # we follow suit since the gemms are tiny.
-        self.wkv = nn.Linear(self.dim, coff * self.head_dim, bias=False)
-        self.wgate = nn.Linear(self.dim, coff * self.head_dim, bias=False)
+        # Compressor weights stay fp32: the reference's Compressor.forward
+        # casts x to fp32 before wkv/wgate, so these linears must accept
+        # fp32 inputs. Set dtype explicitly so this is independent of
+        # torch.get_default_dtype() at construction time.
+        self.wkv = nn.Linear(
+            self.dim, coff * self.head_dim, bias=False, dtype=torch.float32
+        )
+        self.wgate = nn.Linear(
+            self.dim, coff * self.head_dim, bias=False, dtype=torch.float32
+        )
         self.norm = V4RMSNorm(self.head_dim, args.norm_eps)
 
         # Lazily-bound buffers (set by the parent Attention/Indexer):
